ResultsPlot
You can see we have achieved good results. According to Brownian motion Mathematical Engineering the future variations of stock prices are independent and cannot be accurately predict the future stock past data. Keeping that in mind we have achieved a very good results in the model. Open to changes to make the model better and achieve much better results. Pull request for that.